Trying to cleanup pppd.h, breaking out functionality to pppd-private.h
