+++ pppd/chap-md5.c
@@ -36,7 +36,7 @@
 
 #include <stdlib.h>
 #include <string.h>
-#include "pppd.h"
+#include "pppd-private.h"
 #include "chap-new.h"
 #include "chap-md5.h"
 #include "magic.h"
