+++ pppd/plugins/passwordfd.c
@@ -11,15 +11,20 @@
 #include <errno.h>
 #include <string.h>
 #include <unistd.h>
+#include <stdarg.h>
+#include <stdbool.h>
+#include <stdint.h>
+#include <sys/time.h>
 
 #include <pppd/pppd.h>
+#include <pppd/options.h>
 
 char pppd_version[] = PPPD_VERSION;
 
 static int passwdfd = -1;
 static char save_passwd[MAXSECRETLEN];
 
-static option_t options[] = {
+static struct option options[] = {
     { "passwordfd", o_int, &passwdfd,
       "Receive password on this file descriptor" },
     { NULL }
