+++ pppd/sys-solaris.c
@@ -135,7 +135,7 @@
 #include <pcap.h>
 #endif
 
-#include "pppd.h"
+#include "pppd-private.h"
 #include "fsm.h"
 #include "lcp.h"
 #include "ipcp.h"
